backend/requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
alembic==1.12.1
altair==5.5.0
annotated-types==0.7.0
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.9.2
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==24.1.0
Babel==2.15.0
bcrypt==5.0.0
beautifulsoup4==4.12.3
bitarray==3.1.1
bleach==6.1.0
blinker==1.8.2
cachetools==6.1.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.3.2
ckzg==2.0.1
click==8.1.7
colorama==0.4.6
comm==0.2.2
comtypes==1.4.8
contourpy==1.3.2
cors==1.0.1
cryptography==46.0.3
cycler==0.12.1
cytoolz==1.0.1
debugpy==1.8.2
decorator==5.1.1
deepface==0.0.93
defusedxml==0.7.1
distlib==0.3.9
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
et_xmlfile==2.0.0
eth-account==0.13.5
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.6.1
eth-rlp==2.2.0
eth-typing==5.2.0
eth-utils==5.2.0
eth_abi==5.2.0
executing==2.0.1
fastapi==0.104.1
fastjsonschema==2.20.0
filelock==3.16.1
fire==0.7.0
Flask==3.0.3
flask-cors==5.0.1
flatbuffers==25.2.10
fonttools==4.59.0
fqdn==1.5.1
frozenlist==1.5.0
future==1.0.0
gast==0.6.0
gdown==5.2.0
gevent==25.9.1
gitdb==4.0.12
GitPython==3.1.45
google-pasta==0.2.0
greenlet==3.2.4
grpcio==1.71.0
gunicorn==23.0.0
h11==0.14.0
h5py==3.13.0
hexbytes==1.3.0
httpcore==1.0.5
httptools==0.7.1
httpx==0.27.0
idna==3.7
ipykernel==6.29.5
ipython==8.26.0
ipywidgets==8.1.3
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.11
keras==3.9.0
kiwisolver==1.4.8
libclang==18.1.1
lz4==4.4.3
Mako==1.3.10
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.4.1
MouseInfo==0.1.3
mtcnn==1.0.0
multidict==6.1.0
namex==0.0.8
narwhals==2.0.1
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.2.1
notebook_shim==0.2.4
numpy==2.0.1
opencv-python==4.12.0.88
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.14.1
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parsimonious==0.10.0
parso==0.8.4
passlib==1.7.4
pillow==11.3.0
platformdirs==4.2.2
prometheus_client==0.20.0
prompt_toolkit==3.0.47
propcache==0.3.0
protobuf==5.29.3
psutil==6.0.0
psycopg2-binary==2.9.9
pure_eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.5.0
pydantic-settings==2.0.3
pydantic_core==2.14.1
pydeck==0.9.1
PyGetWindow==0.0.9
Pygments==2.18.0
pyjokes==0.6.0
PyMsgBox==1.0.9
pyparsing==3.2.3
pyperclip==1.9.0
pypiwin32==223
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
python-jose==3.3.0
python-json-logger==2.0.7
python-magic-bin==0.4.14
python-multipart==0.0.6
pyttsx3==2.98
pytweening==1.2.0
pytz==2024.2
pyunormalize==16.0.0
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==26.0.3
qtconsole==5.5.2
QtPy==2.4.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.5
requests-file==3.0.1
retina-face==0.0.17
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rlp==4.1.0
rpds-py==0.19.1
rsa==4.9.1
scikit-learn==1.7.1
scipy==1.16.1
seaborn==0.13.2
selenium==4.35.0
Send2Trash==1.8.3
setuptools==72.1.0
sheryanalysis==0.1.0
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SpeechRecognition==3.11.0
SQLAlchemy==2.0.23
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.27.0
streamlit==1.48.0
tenacity==9.1.2
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow_intel==2.18.0
termcolor==2.5.0
terminado==0.18.1
tf_keras==2.18.0
threadpoolctl==3.6.0
tinycss2==1.3.0
tldextract==5.3.0
toml==0.10.2
toolz==1.0.0
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
trio==0.30.0
trio-websocket==0.12.2
types-python-dateutil==2.9.0.20240316
types-requests==2.32.0.20250306
typing_extensions==4.14.1
tzdata==2024.2
uri-template==1.3.0
urllib3==2.5.0
uvicorn==0.24.0
virtualenv==20.28.0
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.2.13
web3==7.8.0
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==13.1
Werkzeug==3.0.3
wheel==0.45.1
widgetsnbextension==4.0.11
wrapt==1.17.2
wsproto==1.2.0
xgboost==3.0.4
yarl==1.18.3
zope.event==6.1
zope.interface==8.0.1
